Changes to build XDP on CLient linux (#8535)

+++ src/runtime_src/xdp/profile/plugin/aie_debug/aie_debug_plugin.cpp
@@ -134,7 +134,7 @@ namespace xdp {
       xrt_core::message::send(severity_level::debug, "XRT", msg.str());
 
       for (auto &tile : tiles) {
-        for (int i = 0; i < Regs.size(); i++) {
+        for (size_t i = 0; i < Regs.size(); i++) {
           op_profile_data.emplace_back(register_data_t{Regs[i] + (tile.col << 25) + (tile.row << 20)});
           counterId++;
         }
@@ -150,7 +150,7 @@ namespace xdp {
     op_size = sizeof(read_register_op_t) + sizeof(register_data_t) * (counterId - 1);
     op = (read_register_op_t*)malloc(op_size);
     op->count = counterId;
-    for (int i = 0; i < op_profile_data.size(); i++)
+    for (size_t i = 0; i < op_profile_data.size(); i++)
       op->data[i] = op_profile_data[i];
 
   #if 0
